+++ plugins/inputs/ping/ping.go
@@ -4,6 +4,8 @@ import (
 	"errors"
 	"fmt"
 	"math"
+	"net"
+	"os/exec"
 	"runtime"
 	"strings"
 	"sync"
@@ -11,6 +13,7 @@ import (
 
 	"github.com/go-ping/ping"
 	"github.com/influxdata/telegraf"
+	"github.com/influxdata/telegraf/internal"
 	"github.com/influxdata/telegraf/plugins/inputs"
 )
 
@@ -26,6 +29,8 @@ type Ping struct {
 	// Pre-calculated interval and timeout
 	calcInterval time.Duration
 	calcTimeout  time.Duration
+
+	sourceAddress string
 
 	Log telegraf.Logger `toml:"-"`
 
@@ -169,6 +174,7 @@ func (p *Ping) nativePing(destination string) (*pingStats, error) {
 		pinger.SetNetwork("ip6")
 	}
 
+	pinger.Source = p.sourceAddress
 	pinger.Interval = p.calcInterval
 	pinger.Timeout = p.calcTimeout
 
@@ -308,12 +314,41 @@ func (p *Ping) Init() error {
 		p.calcTimeout = time.Duration(p.Timeout) * time.Second
 	}
 
+	// Support either an IP address or interface name
+	if p.Interface != "" {
+		if addr := net.ParseIP(p.Interface); addr != nil {
+			p.sourceAddress = p.Interface
+		} else {
+			i, err := net.InterfaceByName(p.Interface)
+			if err != nil {
+				return fmt.Errorf("Failed to get interface: %w", err)
+			}
+			addrs, err := i.Addrs()
+			if err != nil {
+				return fmt.Errorf("Failed to get the address of interface: %w", err)
+			}
+			p.sourceAddress = addrs[0].(*net.IPNet).IP.String()
+		}
+	}
+
 	return nil
+}
+
+func hostPinger(binary string, timeout float64, args ...string) (string, error) {
+	bin, err := exec.LookPath(binary)
+	if err != nil {
+		return "", err
+	}
+	c := exec.Command(bin, args...)
+	out, err := internal.CombinedOutputTimeout(c,
+		time.Second*time.Duration(timeout+5))
+	return string(out), err
 }
 
 func init() {
 	inputs.Add("ping", func() telegraf.Input {
 		p := &Ping{
+			pingHost:     hostPinger,
 			PingInterval: 1.0,
 			Count:        1,
 			Timeout:      1.0,